apps/django_backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
axios
